modelpy/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pillow==10.0.1
scikit-learn==1.3.0
matplotlib==3.8.0
